+++ test/construction.test.js
@@ -131,6 +131,38 @@ describe('construction', function() {
     assert.equal(fn(new Date()), 'Date');
   });
 
+  it('should correctly handle null', function () {
+    var fn = typed({
+      'Object': function (a) {
+        return 'Object';
+      },
+      'null': function (a) {
+        return 'null';
+      },
+      'undefined': function (a) {
+        return 'undefined';
+      }
+    });
+
+    assert.equal(fn(new Object(null)), 'Object');
+    assert.equal(fn(null), 'null');
+    assert.equal(fn(undefined), 'undefined');
+  });
+
+  it('should throw correct error message when passing null from an Object', function() {
+    var signatures = {
+      'Object': function (value) {
+        assert(value instanceof Object);
+        return 'Object';
+      }
+    };
+    var fn = typed(signatures);
+
+    assert.equal(fn({}), 'Object');
+    assert.throws(function () { fn(null) },
+        /TypeError: Unexpected type of argument in function unnamed \(expected: Object, actual: null, index: 0\)/);
+  });
+
   it('should create a new, isolated instance of typed-function', function() {
     var typed1 = typed.create();
     var typed2 = typed.create();